+++ mini-processor/src/main/java/mini/processor/ReducersGenerator.kt
@@ -10,7 +10,6 @@ import javax.lang.model.element.Element
 import javax.lang.model.element.ExecutableElement
 
 object ReducersGenerator {
-
     fun generate(container: TypeSpec.Builder, elements: Set<Element>) {
         val reducers = elements.map { ReducerModel(it) }
             .groupBy { it.containerName }
@@ -18,12 +17,19 @@ object ReducersGenerator {
         val reducerContainerType = Any::class.asTypeName()
         val reducerContainerListType = List::class.asTypeName().parameterizedBy(reducerContainerType)
 
-        val whenBlock = CodeBlock.builder()
+        val newDispatcherFn = FunSpec.builder("newDispatcher")
+            .returns(Dispatcher::class)
+            .addCode(CodeBlock.builder()
+                .addStatement("return Dispatcher(actionTypes)")
+                .build())
+            .build()
+
+        val subscribeSingleWhenBlock = CodeBlock.builder()
             .addStatement("val c = %T()", CompositeCloseable::class)
-            .addStatement("when (container) {").indent()
+            .beginControlFlow("when (container)")
             .apply {
                 reducers.forEach { (containerName, reducerFunctions) ->
-                    addStatement("is %T -> {", containerName).indent()
+                    beginControlFlow("is %T ->", containerName)
                     reducerFunctions.forEach { function ->
                         addStatement("c.add(dispatcher.subscribe<%T>(priority=%L) { container.%N(it) })",
                             function.function.parameters[0].asType(), //Action type
@@ -31,26 +37,29 @@ object ReducersGenerator {
                             function.function.simpleName //Function name
                         )
                     }
-                    unindent().addStatement("}")
+                    endControlFlow()
                 }
             }
             .addStatement("else -> throw IllegalArgumentException(\"Container \$container has no reducers\")")
-            .unindent()
-            .addStatement("}") //Close when
+            .endControlFlow()
             .addStatement("return c")
             .build()
 
-        val registerOneFn = FunSpec.builder("subscribe")
+        val subscribeSingleFn = FunSpec.builder("subscribe")
             .addModifiers(KModifier.PRIVATE)
-            .addParameter("dispatcher", Dispatcher::class)
-            .addParameter("container", reducerContainerType)
+            .addParameters(listOf(
+                ParameterSpec.builder("dispatcher", Dispatcher::class).build(),
+                ParameterSpec.builder("container", reducerContainerType).build()
+            ))
             .returns(Closeable::class)
-            .addCode(whenBlock)
+            .addCode(subscribeSingleWhenBlock)
             .build()
 
-        val registerListFn = FunSpec.builder("subscribe")
-            .addParameter("dispatcher", Dispatcher::class)
-            .addParameter("containers", reducerContainerListType)
+        val subscribeListFn = FunSpec.builder("subscribe")
+            .addParameters(listOf(
+                ParameterSpec.builder("dispatcher", Dispatcher::class).build(),
+                ParameterSpec.builder("containers", reducerContainerListType).build()
+            ))
             .returns(Closeable::class)
             .addStatement("val c = %T()", CompositeCloseable::class)
             .beginControlFlow("containers.forEach { container ->")
@@ -59,23 +68,17 @@ object ReducersGenerator {
             .addStatement("return c")
             .build()
 
-        val initDispatcherFn = FunSpec.builder("newDispatcher")
-            .returns(Dispatcher::class)
-            .addCode(CodeBlock.builder()
-                .addStatement("return Dispatcher(actionTypes)")
-                .build())
-            .build()
-
-        container.addFunction(initDispatcherFn)
-        container.addFunction(registerOneFn)
-        container.addFunction(registerListFn)
-
+        with(container) {
+            addFunction(newDispatcherFn)
+            addFunction(subscribeSingleFn)
+            addFunction(subscribeListFn)
+        }
     }
 }
 
-class ReducerModel(val element: Element) {
+private class ReducerModel(element: Element) {
     val priority = element.getAnnotation(Reducer::class.java).priority
     val function = element as ExecutableElement
-    val container = element.enclosingElement.asType()
+    val container = element.enclosingElement.asType()!!
     val containerName = container.asTypeName()
 }