+++ CHANGELOG.md
@@ -5,7 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- No new features yet
+### Added
+- Add support for incremental annotation processing.
 
 ## [1.2.0] - 2020-02-19
 ### Added